DiscordBot/requirements.txt
discord.py==2.3.1
disco-py
websockets==11.0.3
openai==0.27.8
json-modify==1.0.1
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
httplib2
boto3
alive-progress
tqdm
redis
